refactor: consolidated analytics hooks in one
