--- src/hooks/useFinancialAnalytics.tsx
@@ -1,7 +1,0 @@
-import { getSpendingsQuery } from '@/query/financialAnalytics';
-import { useQuery } from '@tanstack/react-query';
-
-export function useSpendingByCategoriesQuery() {
-  const categoriesSpendingQuery = getSpendingsQuery();
-  return useQuery(categoriesSpendingQuery);
-}
