--- src/hooks/useOverallSpendings.tsx
@@ -1,7 +1,0 @@
-import { overallSpendingQuery } from '@/query/financialAnalytics';
-import { useQuery } from '@tanstack/react-query';
-
-export function useOverallSpending() {
-  const overallSpending = overallSpendingQuery();
-  return useQuery(overallSpending);
-}
