--- src/hooks/useSpendingTrend.tsx
@@ -1,7 +1,0 @@
-import { transactionsTrendQuery } from '@/query/financialAnalytics';
-import { useQuery } from '@tanstack/react-query';
-
-export function useSpendingTrendQuery() {
-  const spendingTrend = transactionsTrendQuery();
-  return useQuery(spendingTrend);
-}
