--- src/hooks/useTopSpendingByCategory.tsx
@@ -1,7 +1,0 @@
-import { getTopSpendingByCategory } from '@/query/financialAnalytics';
-import { useQuery } from '@tanstack/react-query';
-
-export function useTopSpendingByCategory() {
-  const topSpendingByCategoryQuery = getTopSpendingByCategory();
-  return useQuery(topSpendingByCategoryQuery);
-}
